+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - syntax: add few missing cases for trailing commas
 - syntax: add a semicolon at the end of a file if there is a missing newline
 - syntax: add arbitrary precision float, conversion between int and float
+- eval: add option/flag -recomputeempty to recompute empty cache values
 
 ## [reflow0.5.1](https://github.com/grailbio/reflow/releases/tag/reflow0.5.1)